every file under 1000 lines of production code
