+++ atomic-cli/src/commands/mod.rs
@@ -98,9 +98,6 @@ pub mod remote;
 // Phase 4: Identity Management
 pub mod identity;
 
-// Phase 5: Hive Agent Social Platform
-pub mod hive;
-
 // Git Interoperability
 pub mod git;
 
@@ -111,7 +108,6 @@ pub use change::ChangeCmd;
 pub use clone::Clone;
 pub use diff::Diff;
 pub use git::Git;
-pub use hive::Hive;
 pub use identity::Identity;
 pub use init::Init;
 pub use insert::Insert;
